# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: CTWall Backend API
  version: 1.0.0
  description: |
    API for ChainThreatWall (CTWall) - SBOM Management and Security Analysis System.
    Adheres to OWASP ASVS Level 2 Security Requirements.
    Uses RFC 7807 for error handling.
servers:
  - url: https://api.ctwall.internal/api/v1
    description: Production Server
  - url: http://localhost:8080/api/v1
    description: Local Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        API token for user/service accounts/CLI.
        Use: `Authorization: Bearer <API_TOKEN>`.
    CookieAuth:
      type: apiKey
      in: cookie
      name: __Host-CTWALL_SESSION
      description: |
        Session JWT stored in HttpOnly cookie for GUI users.
    RefreshCookieAuth:
      type: apiKey
      in: cookie
      name: __Secure-CTWALL_REFRESH
      description: |
        Refresh token stored in HttpOnly cookie for GUI session renewal.
  schemas:
...
    Error:
      type: object
      description: RFC 7807 Problem Details
      properties:
        type:
          type: string
          maxLength: 120
          format: uri
          example: "about:blank"
        title:
          type: string
          maxLength: 120
          example: "Invalid Request"
        status:
          type: integer
          example: 400
        detail:
          type: string
          maxLength: 120
          example: "Field 'sbom_file' is missing."
        instance:
          type: string
          maxLength: 120
          example: "/api/v1/ingest"
...
    AlertGroup:
      type: object
      required: [id, projectId, severity, category, type, status, groupKey, title, occurrences, firstSeenAt, lastSeenAt, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        severity: { type: string, enum: [INFO, WARN, ERROR] }
        category: { type: string }
        type: { type: string }
        status: { type: string, enum: [OPEN, ACKNOWLEDGED, CLOSED] }
        groupKey: { type: string }
        title: { type: string }
        entityRef: { type: string, nullable: true }
        occurrences: { type: integer, minimum: 0 }
        firstSeenAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        lastNotifiedAt: { type: string, format: date-time, nullable: true }
        acknowledgedAt: { type: string, format: date-time, nullable: true }
        acknowledgedBy: { type: string, format: uuid, nullable: true }
        closedAt: { type: string, format: date-time, nullable: true }
        closedBy: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AlertOccurrence:
      type: object
      required: [id, projectId, groupId, severity, category, type, title, occurredAt, details, createdAt]
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        groupId: { type: string, format: uuid }
        severity: { type: string, enum: [INFO, WARN, ERROR] }
        category: { type: string }
        type: { type: string }
        title: { type: string }
        occurredAt: { type: string, format: date-time }
        productId: { type: string, format: uuid, nullable: true }
        scopeId: { type: string, format: uuid, nullable: true }
        testId: { type: string, format: uuid, nullable: true }
        entityRef: { type: string, nullable: true }
        details: {}
        createdAt: { type: string, format: date-time }
    AlertGroupsListResponse:
      type: object
      required: [items, page, pageSize, total, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AlertGroup"
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
    AlertOccurrencesListResponse:
      type: object
      required: [items, page, pageSize, total, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AlertOccurrence"
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /alert-groups:
    get:
      summary: List alert groups (batched, project-scoped)
      operationId: listAlertGroups
      parameters:
        - name: severity
          in: query
          required: false
          schema:
            type: string
            maxLength: 120
          description: Comma-separated list (INFO,WARN,ERROR).
        - name: category
          in: query
          required: false
          schema:
            type: string
            maxLength: 400
          description: Comma-separated list of categories (eventmeta.Category).
        - name: type
          in: query
          required: false
          schema:
            type: string
            maxLength: 400
          description: Comma-separated list of low-cardinality alert types (e.g. malware.detected).
        - name: status
          in: query
          required: false
          schema:
            type: string
            maxLength: 64
          description: Comma-separated list of statuses (OPEN,ACKNOWLEDGED,CLOSED).
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 128
          description: Optional search string (title/entity/type/category).
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Optional RFC3339 start timestamp (last_seen_at).
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Optional RFC3339 end timestamp (last_seen_at).
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Alert groups list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertGroupsListResponse"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /alert-groups/{id}:
    get:
      summary: Get alert group by id (project-scoped)
      operationId: getAlertGroup
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alert group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertGroup"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /alert-groups/{id}/acknowledge:
    post:
      summary: Acknowledge an alert group (admin, non-malware)
      operationId: acknowledgeAlertGroup
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Acknowledged
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '409':
          description: Conflict (malware groups managed via Explorer triage)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /alert-groups/{id}/close:
    post:
      summary: Close an alert group (admin, non-malware)
      operationId: closeAlertGroup
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Closed
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '409':
          description: Conflict (malware groups managed via Explorer triage)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /alert-occurrences:
    get:
      summary: List alert occurrences (append-only, project-scoped)
      operationId: listAlertOccurrences
      parameters:
        - name: groupId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Optional group id filter.
        - name: severity
          in: query
          required: false
          schema:
            type: string
            maxLength: 120
          description: Comma-separated list (INFO,WARN,ERROR) (filters by group severity).
        - name: category
          in: query
          required: false
          schema:
            type: string
            maxLength: 400
          description: Comma-separated list of categories (filters by group category).
        - name: type
          in: query
          required: false
          schema:
            type: string
            maxLength: 400
          description: Comma-separated list of types (filters by group type).
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 128
          description: Optional search string (entity/title/type).
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Optional RFC3339 start timestamp (occurred_at).
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Optional RFC3339 end timestamp (occurred_at).
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Alert occurrences list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertOccurrencesListResponse"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...